fix: 폰트 수정 배치수정
