+++ src/components/common/footer.tsx
@@ -1,171 +1,96 @@
+import Image from "next/image";
+import { useRouter } from "next/router";
 import { cn } from "src/core/function/cn";
 
 const Footer = ({ className }: { className: string }) => {
+  const router = useRouter();
   return (
-    <footer className={cn("body-font text-gray-600", className)}>
-      <div className="container mx-auto flex flex-col flex-wrap px-5 py-24 md:flex-row md:flex-nowrap md:items-center lg:items-start">
-        <div className="mx-auto w-auto flex-shrink-0 text-center md:mx-0 md:text-left">
-          <a className="title-font flex items-center justify-center font-medium text-gray-900 md:justify-start">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              fill="none"
-              stroke="currentColor"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="2"
-              className="h-10 w-10 rounded-full bg-indigo-500 p-2 text-white"
-              viewBox="0 0 24 24"
-            >
-              <path d="M12 2L2 7l10 5 10-5-10-5zM2 17l10 5 10-5M2 12l10 5 10-5"></path>
-            </svg>
-            <span className="ml-3 text-xl">Wavle</span>
-          </a>
-          <p className="mt-2 text-sm text-gray-500">Air plant banjo lyft occupy retro adaptogen indego</p>
+    <div className={cn("fixed bottom-0 w-full bg-[#ffffff] pt-5", className)}>
+      <div className="flex justify-evenly bg-[#ffffff] text-[0.625rem]/[1.5rem] font-medium text-[#4a5360]">
+        <div className="flex flex-1 cursor-pointer flex-col items-center justify-center">
+          <div
+            className="relative h-[1.5rem] w-[1.5rem]"
+            onClick={() => {
+              router.push("/");
+            }}
+          >
+            <Image
+              src={`/images/assets/common/${router.asPath.startsWith("/") ? "home-active" : "home"}.svg`}
+              alt="home"
+              fill
+              className="object-cover"
+            />
+          </div>
+          <span>모임</span>
         </div>
-        <div className="-mb-10 mt-10 flex flex-grow flex-wrap text-center md:mt-0 md:pl-20 md:text-left">
-          <div className="w-full px-4 md:w-1/2 lg:w-1/4">
-            <h2 className="title-font mb-3 text-sm font-medium tracking-widest text-gray-900">CATEGORIES</h2>
-            <nav className="mb-10 list-none">
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">First Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Second Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Third Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Fourth Link</a>
-              </li>
-            </nav>
+        <div className="flex flex-1 cursor-pointer flex-col items-center justify-center">
+          <div
+            className="relative h-[1.5rem] w-[1.5rem]"
+            onClick={() => {
+              router.push("/");
+            }}
+          >
+            <Image
+              src={`/images/assets/common/${!router.asPath.startsWith("/") ? "article-active" : "article"}.svg`}
+              fill
+              className="object-cover"
+              alt="article"
+            />
           </div>
-          <div className="w-full px-4 md:w-1/2 lg:w-1/4">
-            <h2 className="title-font mb-3 text-sm font-medium tracking-widest text-gray-900">CATEGORIES</h2>
-            <nav className="mb-10 list-none">
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">First Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Second Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Third Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Fourth Link</a>
-              </li>
-            </nav>
+          <span>아티클</span>
+        </div>
+        <div className="flex flex-1 cursor-pointer flex-col items-center justify-center">
+          <div
+            className="relative h-[1.5rem] w-[1.5rem]"
+            onClick={() => {
+              router.push("/");
+            }}
+          >
+            <Image
+              src={`/images/assets/common/${
+                !router.asPath.startsWith("/") ? "around-location-active" : "around-location"
+              }.svg`}
+              fill
+              className="object-cover"
+              alt="around-location"
+            />
           </div>
-          <div className="w-full px-4 md:w-1/2 lg:w-1/4">
-            <h2 className="title-font mb-3 text-sm font-medium tracking-widest text-gray-900">CATEGORIES</h2>
-            <nav className="mb-10 list-none">
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">First Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Second Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Third Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Fourth Link</a>
-              </li>
-            </nav>
+          <span>주변 탐색</span>
+        </div>
+        <div className="flex flex-1 cursor-pointer flex-col items-center justify-center">
+          <div
+            className="relative h-[1.5rem] w-[1.5rem]"
+            onClick={() => {
+              router.push("/");
+            }}
+          >
+            <Image
+              src={`/images/assets/common/${!router.asPath.startsWith("/") ? "chatting-active" : "chatting"}.svg`}
+              fill
+              className="object-cover"
+              alt="chatting"
+            />
           </div>
-          <div className="w-full px-4 md:w-1/2 lg:w-1/4">
-            <h2 className="title-font mb-3 text-sm font-medium tracking-widest text-gray-900">CATEGORIES</h2>
-            <nav className="mb-10 list-none">
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">First Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Second Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Third Link</a>
-              </li>
-              <li>
-                <a className="text-gray-600 hover:text-gray-800">Fourth Link</a>
-              </li>
-            </nav>
+          <span>채팅</span>
+        </div>
+        <div className="flex flex-1 cursor-pointer flex-col items-center justify-center">
+          <div
+            className="relative h-[1.5rem] w-[1.5rem]"
+            onClick={() => {
+              router.push("/");
+            }}
+          >
+            <Image
+              src={`/images/assets/common/${!router.asPath.startsWith("/") ? "my-page-active" : "my-page"}.svg`}
+              fill
+              className="object-cover"
+              alt="my-page"
+            />
           </div>
+          <span>마이페이지</span>
         </div>
       </div>
-      <div className="bg-gray-100">
-        <div className="container mx-auto flex flex-col flex-wrap items-center px-5 py-4 sm:flex-row">
-          <p className="text-center text-sm text-gray-500 sm:text-left">
-            © 2023 Wavle —
-            <a
-              href="https://twitter.com/knyttneve"
-              rel="noopener noreferrer"
-              className="ml-1 text-gray-600"
-              target="_blank"
-            >
-              @Jiheum
-            </a>
-          </p>
-          <span className="mt-2 inline-flex justify-center sm:ml-auto sm:mt-0 sm:justify-start">
-            <a className="text-gray-500">
-              <svg
-                fill="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth="2"
-                className="h-15 w-15"
-                viewBox="0 0 24 24"
-              >
-                <path d="M18 2h-3a5 5 0 00-5 5v3H7v4h3v8h4v-8h3l1-4h-4V7a1 1 0 011-1h3z"></path>
-              </svg>
-            </a>
-            <a className="ml-3 text-gray-500">
-              <svg
-                fill="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth="2"
-                className="h-15 w-15"
-                viewBox="0 0 24 24"
-              >
-                <path d="M23 3a10.9 10.9 0 01-3.14 1.53 4.48 4.48 0 00-7.86 3v1A10.66 10.66 0 013 4s-4 9 5 13a11.64 11.64 0 01-7 2c9 5 20 0 20-11.5a4.5 4.5 0 00-.08-.83A7.72 7.72 0 0023 3z"></path>
-              </svg>
-            </a>
-            <a className="ml-3 text-gray-500">
-              <svg
-                fill="none"
-                stroke="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth="2"
-                className="h-15 w-15"
-                viewBox="0 0 24 24"
-              >
-                <rect width="20" height="20" x="2" y="2" rx="5" ry="5"></rect>
-                <path d="M16 11.37A4 4 0 1112.63 8 4 4 0 0116 11.37zm1.5-4.87h.01"></path>
-              </svg>
-            </a>
-            <a className="ml-3 text-gray-500">
-              <svg
-                fill="currentColor"
-                stroke="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth="0"
-                className="h-15 w-15"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  stroke="none"
-                  d="M16 8a6 6 0 016 6v7h-4v-7a2 2 0 00-2-2 2 2 0 00-2 2v7h-4v-7a6 6 0 016-6zM2 9h4v12H2z"
-                ></path>
-                <circle cx="4" cy="4" r="2" stroke="none"></circle>
-              </svg>
-            </a>
-          </span>
-        </div>
-      </div>
-    </footer>
+    </div>
   );
 };
